fix bug and refactor code tread
